2023/01/10
1. 그동안 junit mockito를 잘못 사용하고 있었다는 것을 알게되었다...

2. controller tdd를 하던 도중 계속 문제가 생겨 해결하기 위해 반나절 이상을 삽질했는데
결국에는 내가 잘못 이해하고 있었다는 사실을 알았다.

3. 오늘 작업은 여기까지 하고 운동, 식사 후 프로그래머스로 코딩 테스트 공부를 할 예정이다.

4. 2023/01/11 내일은 남은 controller를 tdd로 전부 구현할 예정이며 작업이 빠르게 끝난다면 이미지 업로드, 동영상 업로드 등의 service, controller 등을 작업.

+++ src/main/java/Baetube_backEnd/service/contents/ContentsDeleteService.java
@@ -1,22 +1,27 @@
 package Baetube_backEnd.service.contents;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.transaction.annotation.Transactional;
 
 import Baetube_backEnd.dto.Contents;
+import Baetube_backEnd.exception.NullContentsException;
 import Baetube_backEnd.mapper.ContentsMapper;
 
 public class ContentsDeleteService
 {
 	@Autowired
 	private ContentsMapper contentsMapper;
-
-	public void setContentsMapper(ContentsMapper contentsMapper)
-	{
-		this.contentsMapper = contentsMapper;
-	}
 	
+	@Transactional
 	public boolean deleteContents(Contents request)
 	{
+		Contents contents = contentsMapper.selectByContentsId(request.getContentsId());
+		
+		if(contents == null)
+		{
+			throw new NullContentsException();
+		}
+		
 		contentsMapper.delete(request.getContentsId(), request.getType());
 		
 		return true;
